Fixed more warnings
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@933 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Tools/Utils/StripifierT.hh
@@ -92,7 +92,7 @@ public:
   ~StripifierT();
 
   /// Compute triangle strips, returns number of strips
-  unsigned int stripify();
+  size_t stripify();
 
   /// delete all strips
   void clear() { Strips().swap(strips_); }
